+++ src/content/index.ts
@@ -1,3 +1,5 @@
 import { init as initHideExplore } from './hide-explore';
+import { init as initHideReels } from './hide-reels';
 
 initHideExplore();
+initHideReels();